refactor: moved PortMapping from package org.eclipse.jkube.kit.build.api.model to org.eclipse.jkube.kit.common
