+++ jkube-kit/common/src/main/java/org/eclipse/jkube/kit/common/PortMapping.java
@@ -11,7 +11,7 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.jkube.kit.build.api.model;
+package org.eclipse.jkube.kit.common;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
